+++ src/pages/gallery/questions/question-intro.tsx
@@ -1,8 +1,8 @@
 import { GetStaticProps } from "next";
 import path from "path";
 
-import { getPropsForMarkdownFile } from "../../lib/markdown";
-import { MarkdownPage, MarkdownPageProps } from "../../lib/markdown-page";
+import { getPropsForMarkdownFile } from "../../../lib/markdown";
+import { MarkdownPage, MarkdownPageProps } from "../../../lib/markdown-page";
 
 export default MarkdownPage;
 
@@ -13,6 +13,7 @@ export const getStaticProps: GetStaticProps<MarkdownPageProps> = async () => {
         process.cwd(),
         "src",
         "pages",
+        "gallery",
         "questions",
         "question-intro.md"
       )
